Lesson delete requirements added

+++ employee/views.py
@@ -10,6 +10,8 @@ from manager.models import Request
 from bank.models import StudentAccount, StudentLessonBill
 import re
 import datetime
+from datetime import timedelta
+from django.utils import timezone
 
 def validation_employee(func):
 
@@ -479,7 +481,8 @@ def lesson_delete(request, id):
     lesson = context['lesson'] = Lesson.objects.get(pk = id)
 
     if (request.method == 'POST'):
-        # TODO: Check if there is less than N hours before class
+        if timezone.now() > lesson.start or lesson.start - timezone.now() < timedelta(days=1):
+            return render(request, 'employee/403.jinja', context)
         lesson.delete()
         return redirect('employee_lessons')
